Fix deposit/withdraw flow
Deposit (`generateLink` method) has a workaround to be able to follow
with the withdrawal process. But it's calling `deposit` after `transferAndCall`
which is not correct, as this should be an atomic operation.

On the other hand, for the withdraw there's the `canWithdraw` method which
will give an error if the _withdraw_ operation cannot be made, or will
return an object with the `amount` and the `sender`. That information is
temporarily displayed by the `John Doe` string, but is intended to be used
for the information dialog.